pyproject.toml
[project]
name = "langchaindev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.15",
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.68",
    "python-dotenv>=1.0.1",
]
